# openapi/gateway.openapi.json
{
  "components": {
    "schemas": {
      "AgentRequest": {
        "description": "Input to ``POST /v1/agent`` (Wave 7).",
        "example": {
          "max_steps": 4,
          "question": "summarise the README",
          "session_id": "user-42"
        },
        "properties": {
          "max_steps": {
            "anyOf": [
              {
                "maximum": 20.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Per-request override of the AGENT_MAX_STEPS budget.",
            "title": "Max Steps"
          },
          "question": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Question",
            "type": "string"
          },
          "session_id": {
            "anyOf": [
              {
                "maxLength": 128,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "If set, the agent loads the session's prior turns and persists the new pair.",
            "title": "Session Id"
          },
          "use_chat_history": {
            "default": true,
            "description": "If false, session_id is ignored (stateless turn).",
            "title": "Use Chat History",
            "type": "boolean"
          }
        },
        "required": [
          "question"
        ],
        "title": "AgentRequest",
        "type": "object"
      },
      "AgentResponse": {
        "description": "Result of ``POST /v1/agent`` (Wave 7).",
        "properties": {
          "answer": {
            "title": "Answer",
            "type": "string"
          },
          "route": {
            "description": "Final routing label (``kb`` or ``chitchat``).",
            "title": "Route",
            "type": "string"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "sources_used": {
            "default": 0,
            "minimum": 0.0,
            "title": "Sources Used",
            "type": "integer"
          },
          "steps_used": {
            "minimum": 0.0,
            "title": "Steps Used",
            "type": "integer"
          },
          "tool_calls": {
            "items": {
              "$ref": "#/components/schemas/AgentToolCallEntry"
            },
            "title": "Tool Calls",
            "type": "array"
          },
          "trace": {
            "description": "Ordered list of graph node visits for this invocation.",
            "items": {
              "type": "string"
            },
            "title": "Trace",
            "type": "array"
          }
        },
        "required": [
          "answer",
          "route",
          "steps_used"
        ],
        "title": "AgentResponse",
        "type": "object"
      },
      "AgentToolCallEntry": {
        "description": "One tool invocation, suitable for inclusion in :class:`AgentResponse`.",
        "properties": {
          "duration_ms": {
            "minimum": 0.0,
            "title": "Duration Ms",
            "type": "number"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "input": {
            "additionalProperties": true,
            "title": "Input",
            "type": "object"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "output_summary": {
            "description": "Tool output reduced to a string for transport (e.g. retriever returns '<n> docs', calc returns the number, web_search returns '<n> hits'). Full payloads stay inside the agent state for tracing.",
            "title": "Output Summary",
            "type": "string"
          }
        },
        "required": [
          "name",
          "output_summary",
          "duration_ms"
        ],
        "title": "AgentToolCallEntry",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "QueryRequest": {
        "example": {
          "question": "What does foo return?",
          "session_id": "user-42",
          "top_k": 5
        },
        "properties": {
          "question": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Question",
            "type": "string"
          },
          "session_id": {
            "anyOf": [
              {
                "maxLength": 128,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "If set, prior turns in this session are sent to the model and updated after each reply.",
            "title": "Session Id"
          },
          "top_k": {
            "default": 5,
            "description": "Number of KB chunks to merge into context",
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Top K",
            "type": "integer"
          },
          "use_chat_history": {
            "default": true,
            "description": "If false, session_id is ignored (stateless turn).",
            "title": "Use Chat History",
            "type": "boolean"
          }
        },
        "required": [
          "question"
        ],
        "title": "QueryRequest",
        "type": "object"
      },
      "QueryResponse": {
        "properties": {
          "answer": {
            "title": "Answer",
            "type": "string"
          },
          "context_preview": {
            "description": "Short preview of merged retrieval context",
            "title": "Context Preview",
            "type": "string"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Echo when conversation state was used",
            "title": "Session Id"
          },
          "sources_used": {
            "default": 0,
            "description": "How many KB chunks were merged into context",
            "title": "Sources Used",
            "type": "integer"
          }
        },
        "required": [
          "answer",
          "context_preview"
        ],
        "title": "QueryResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "BFF orchestrating ingest and RAG microservices",
    "title": "sorakAi Gateway",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/agent": {
      "post": {
        "operationId": "proxy_agent_v1_agent_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Proxy Agent",
        "tags": [
          "agent"
        ]
      }
    },
    "/v1/agent/stream": {
      "post": {
        "operationId": "proxy_agent_stream_v1_agent_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Proxy Agent Stream",
        "tags": [
          "agent"
        ]
      }
    },
...
    "/v1/query": {
      "post": {
        "operationId": "proxy_query_v1_query_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Proxy Query",
        "tags": [
          "rag"
        ]
      }
    },
    "/v1/query/stream": {
      "post": {
        "operationId": "proxy_query_stream_v1_query_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Proxy Query Stream",
        "tags": [
          "rag"
        ]
      }
    }
  },
...
}